Move the menu into an own module
